requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv

